Wait for inputHandler to complete, too.
I've noticed that occasionally, commands whose only job is to send a PUB
message to another channel do not execute, especially if multiple messages
triggering them arrive in quick succession. It appears that due to all the
goroutines floating around, it is possible for os.exec.Wait to complete before
inputHandler actually gets to do anything to the pipe -- and Wait does not
appear to wait for the pipe to be closed on both ends too, so commands can get
lost entirely or even fail because the pipe was already closed.

Waiting for the inputHandler to complete appears to fix the issue.

Concurrency, fun for the whole family.